Apply namespacing script
This is a shim for allowing Bugsnag+KSCrash to play nicer with other
KSCrash-using libraries in a user's app. It runs the namespacer script:

    ruby Scripts/namespacer.rb

Which moves all classes into a "Bugsnag" namespace. In case of a
conflict, this commit can be recreated by rerunning the script on
the contents of the previous commit.